K8SPS-179: Bootstrap Group Replication
This commit moves group replication configuration logic from operator to
pods. From now, every pod will configure itself for group replication
and join to the cluster.

Bootstrapping the group is only done in pod-0. The need for bootstrap is
determined by the peer count of pod-0: if there's only one peer, we
bootstrap. Having only one peer can happen in two cases:

1. Cluster is deployed for the first time
2. All pods are deleted from cluster and they're starting again (full
   cluster crash)

These two cases need bootstrapping from pod-0. Currently we don't check
GTIDs before bootstrapping. This may be especially important for
rebooting from full cluster crash to not lose any data. Current logic
always bootstraps from pod-0. We need to address this before GA.
