+++ cmd/bootstrap/main.go
@@ -1,23 +1,11 @@
 package main
 
 import (
-	"fmt"
-	"io/ioutil"
 	"log"
-	"net"
 	"os"
 	"path/filepath"
-	"strings"
-	"time"
 
-	"github.com/pkg/errors"
-	"github.com/sjmudd/stopwatch"
-	"k8s.io/apimachinery/pkg/util/sets"
-
-	apiv1alpha1 "github.com/percona/percona-server-mysql-operator/api/v1alpha1"
-	"github.com/percona/percona-server-mysql-operator/pkg/k8s"
 	"github.com/percona/percona-server-mysql-operator/pkg/mysql"
-	"github.com/percona/percona-server-mysql-operator/pkg/replicator"
 )
 
 func main() {
@@ -28,335 +16,17 @@ func main() {
 	defer f.Close()
 	log.SetOutput(f)
 
-	if err := bootstrap(); err != nil {
-		log.Fatalf("bootstrap failed: %v", err)
-	}
-}
-
-func bootstrap() error {
-	timer := stopwatch.NewNamedStopwatch()
-	err := timer.AddMany([]string{"clone", "total"})
-	if err != nil {
-		return errors.Wrap(err, "add timers")
-	}
-	timer.Start("total")
-
-	defer func() {
-		timer.Stop("total")
-		log.Printf("bootstrap finished in %f seconds", timer.ElapsedSeconds("total"))
-	}()
-
-	svc := os.Getenv("SERVICE_NAME_UNREADY")
-	mysqlSvc := os.Getenv("SERVICE_NAME")
-	peers, err := lookup(svc)
-	if err != nil {
-		return errors.Wrap(err, "lookup")
-	}
-	log.Printf("Peers: %v", peers.List())
-
-	exists, err := lockExists()
-	if err != nil {
-		return errors.Wrap(err, "lock file check")
-	}
-	if exists {
-		log.Printf("Waiting for bootstrap.lock to be deleted")
-		if err = waitLockRemoval(); err != nil {
-			return errors.Wrap(err, "wait lock removal")
+	clusterType := os.Getenv("CLUSTER_TYPE")
+	switch clusterType {
+	case "group-replication":
+		if err := bootstrapGroupReplication(); err != nil {
+			log.Fatalf("bootstrap failed: %v", err)
 		}
-	}
-	primary, replicas, err := getTopology(peers)
-	if err != nil {
-		return errors.Wrap(err, "select donor")
-	}
-	log.Printf("Primary: %s Replicas: %v", primary, replicas)
-
-	fqdn, err := getFQDN(mysqlSvc)
-	if err != nil {
-		return errors.Wrap(err, "get FQDN")
-	}
-	log.Printf("FQDN: %s", fqdn)
-
-	podHostname, err := os.Hostname()
-	if err != nil {
-		return errors.Wrap(err, "get hostname")
-	}
-
-	podIp, err := getPodIP(podHostname)
-	if err != nil {
-		return errors.Wrap(err, "get pod IP")
-	}
-	log.Printf("PodIP: %s", podIp)
-
-	primaryIp, err := getPodIP(primary)
-	if err != nil {
-		return errors.Wrap(err, "get primary IP")
-	}
-	log.Printf("PrimaryIP: %s", primaryIp)
-
-	donor, err := selectDonor(fqdn, primary, replicas)
-	if err != nil {
-		return errors.Wrap(err, "select donor")
-	}
-	log.Printf("Donor: %s", donor)
-
-	if donor == "" || donor == fqdn || primary == fqdn || primaryIp == podIp {
-		return nil
-	}
-
-	log.Printf("Opening connection to %s", podIp)
-	operatorPass, err := getSecret(apiv1alpha1.UserOperator)
-	if err != nil {
-		return errors.Wrapf(err, "get %s password", apiv1alpha1.UserOperator)
-	}
-
-	db, err := replicator.NewReplicator("operator", operatorPass, podIp, mysql.DefaultAdminPort)
-	if err != nil {
-		return errors.Wrap(err, "connect to db")
-	}
-	defer db.Close()
-
-	cloneLock := filepath.Join(mysql.DataMountPath, "clone.lock")
-	requireClone, err := isCloneRequired(cloneLock)
-	if err != nil {
-		return errors.Wrap(err, "check if clone is required")
-	}
-
-	log.Printf("Clone required: %t", requireClone)
-	if requireClone {
-		log.Println("Checking if a clone in progress")
-		inProgress, err := db.CloneInProgress()
-		if err != nil {
-			return errors.Wrap(err, "check if a clone in progress")
+	case "async":
+		if err := bootstrapAsyncReplication(); err != nil {
+			log.Fatalf("bootstrap failed: %v", err)
 		}
-
-		log.Printf("Clone in progress: %t", inProgress)
-		if inProgress {
-			return nil
-		}
-
-		timer.Start("clone")
-		log.Printf("Cloning from %s", donor)
-		err = db.Clone(donor, "operator", operatorPass, mysql.DefaultAdminPort)
-		timer.Stop("clone")
-		if err != nil && !errors.Is(err, replicator.ErrRestartAfterClone) {
-			return errors.Wrapf(err, "clone from donor %s", donor)
-		}
-
-		err = createCloneLock(cloneLock)
-		if err != nil {
-			return errors.Wrap(err, "create clone lock")
-		}
-
-		log.Println("Clone finished. Restarting container...")
-
-		// We return with 1 to restart container
-		os.Exit(1)
+	default:
+		log.Fatalf("Invalid cluster type: %v", clusterType)
 	}
-
-	if !requireClone {
-		if err := deleteCloneLock(cloneLock); err != nil {
-			return errors.Wrap(err, "delete clone lock")
-		}
-	}
-
-	rStatus, _, err := db.ReplicationStatus()
-	if err != nil {
-		return errors.Wrap(err, "check replication status")
-	}
-
-	if rStatus == replicator.ReplicationStatusNotInitiated {
-		log.Println("configuring replication")
-
-		replicaPass, err := getSecret(apiv1alpha1.UserReplication)
-		if err != nil {
-			return errors.Wrapf(err, "get %s password", apiv1alpha1.UserReplication)
-		}
-
-		if err := db.StopReplication(); err != nil {
-			return errors.Wrap(err, "stop replication")
-		}
-
-		if err := db.StartReplication(primary, replicaPass, mysql.DefaultPort); err != nil {
-			return errors.Wrap(err, "start replication")
-		}
-	}
-
-	if err := db.EnableSuperReadonly(); err != nil {
-		return errors.Wrap(err, "enable super read only")
-	}
-
-	return nil
-}
-
-func getFQDN(svcName string) (string, error) {
-	hostname, err := os.Hostname()
-	if err != nil {
-		return "", errors.Wrap(err, "get hostname")
-	}
-
-	namespace, err := k8s.DefaultAPINamespace()
-	if err != nil {
-		return "", errors.Wrap(err, "get namespace")
-	}
-
-	return fmt.Sprintf("%s.%s.%s", hostname, svcName, namespace), nil
-}
-
-func getSecret(username apiv1alpha1.SystemUser) (string, error) {
-	path := filepath.Join(mysql.CredsMountPath, string(username))
-	sBytes, err := ioutil.ReadFile(path)
-	if err != nil {
-		return "", errors.Wrapf(err, "read %s", path)
-	}
-
-	return strings.TrimSpace(string(sBytes)), nil
-}
-
-func getPodIP(hostname string) (string, error) {
-	addrs, err := net.LookupHost(hostname)
-	if err != nil {
-		return "", errors.Wrapf(err, "lookup %s", hostname)
-	}
-	log.Println("lookup", hostname, addrs)
-
-	return addrs[0], nil
-}
-
-func lookup(svcName string) (sets.String, error) {
-	endpoints := sets.NewString()
-	_, srvRecords, err := net.LookupSRV("", "", svcName)
-	if err != nil {
-		return endpoints, err
-	}
-	for _, srvRecord := range srvRecords {
-		// The SRV records have the pattern $HOSTNAME.$SERVICE.$.NAMESPACE.svc.$CLUSTER_DNS_SUFFIX
-		// We only want $HOSTNAME.$SERVICE.$NAMESPACE because in the `selectDonor` function we
-		// compare the list generated here with the output of the `getFQDN` function
-		srv := strings.Split(srvRecord.Target, ".")
-		ep := strings.Join(srv[:3], ".")
-		endpoints.Insert(ep)
-	}
-	return endpoints, nil
-}
-
-func getTopology(peers sets.String) (string, []string, error) {
-	replicas := sets.NewString()
-	primary := ""
-
-	operatorPass, err := getSecret(apiv1alpha1.UserOperator)
-	if err != nil {
-		return "", nil, errors.Wrapf(err, "get %s password", apiv1alpha1.UserOperator)
-	}
-
-	for _, peer := range peers.List() {
-		db, err := replicator.NewReplicator("operator", operatorPass, peer, mysql.DefaultAdminPort)
-		if err != nil {
-			return "", nil, errors.Wrapf(err, "connect to %s", peer)
-		}
-		defer db.Close()
-
-		status, source, err := db.ReplicationStatus()
-		if err != nil {
-			return "", nil, errors.Wrap(err, "check replication status")
-		}
-
-		replicaHost, err := db.ReportHost()
-		if err != nil {
-			return "", nil, errors.Wrap(err, "get report_host")
-		}
-		if replicaHost == "" {
-			continue
-		}
-		replicas.Insert(replicaHost)
-
-		if status == replicator.ReplicationStatusActive {
-			primary = source
-		}
-	}
-
-	if primary == "" && peers.Len() == 1 {
-		primary = peers.List()[0]
-	} else if primary == "" {
-		primary = replicas.List()[0]
-	}
-
-	if replicas.Len() > 0 {
-		replicas.Delete(primary)
-	}
-
-	return primary, replicas.List(), nil
-}
-
-func selectDonor(fqdn, primary string, replicas []string) (string, error) {
-	donor := ""
-
-	operatorPass, err := getSecret(apiv1alpha1.UserOperator)
-	if err != nil {
-		return "", errors.Wrapf(err, "get %s password", apiv1alpha1.UserOperator)
-	}
-
-	for _, replica := range replicas {
-		db, err := replicator.NewReplicator("operator", operatorPass, replica, mysql.DefaultAdminPort)
-		if err != nil {
-			continue
-		}
-		db.Close()
-
-		if fqdn != replica {
-			donor = replica
-			break
-		}
-	}
-
-	if donor == "" && fqdn != primary {
-		donor = primary
-	}
-
-	return donor, nil
-}
-
-func isCloneRequired(file string) (bool, error) {
-	_, err := os.Stat(file)
-	if err != nil {
-		if os.IsNotExist(err) {
-			return true, nil
-		}
-		return false, errors.Wrapf(err, "stat %s", file)
-	}
-
-	return false, nil
-}
-
-func createCloneLock(file string) error {
-	_, err := os.Create(file)
-	return errors.Wrapf(err, "create %s", file)
-}
-
-func deleteCloneLock(file string) error {
-	err := os.Remove(file)
-	return errors.Wrapf(err, "remove %s", file)
-}
-
-func waitLockRemoval() error {
-	for {
-		exists, err := lockExists()
-		if err != nil {
-			return err
-		}
-		time.Sleep(time.Second)
-		if !exists {
-			return nil
-		}
-	}
-}
-
-func lockExists() (bool, error) {
-	_, err := os.Stat("/var/lib/mysql/bootstrap.lock")
-	if err != nil {
-		if os.IsNotExist(err) {
-			return false, nil
-		}
-		return false, errors.Wrap(err, "os stat")
-	}
-	return true, nil
 }
