+++ cmd/bootstrap/utils.go
@@ -1,0 +1,94 @@
+package main
+
+import (
+	"fmt"
+	"io/ioutil"
+	"log"
+	"net"
+	"os"
+	"path/filepath"
+	"strings"
+	"time"
+
+	"github.com/pkg/errors"
+	"k8s.io/apimachinery/pkg/util/sets"
+
+	apiv1alpha1 "github.com/percona/percona-server-mysql-operator/api/v1alpha1"
+	"github.com/percona/percona-server-mysql-operator/pkg/k8s"
+	"github.com/percona/percona-server-mysql-operator/pkg/mysql"
+)
+
+func getFQDN(svcName string) (string, error) {
+	hostname, err := os.Hostname()
+	if err != nil {
+		return "", errors.Wrap(err, "get hostname")
+	}
+
+	namespace, err := k8s.DefaultAPINamespace()
+	if err != nil {
+		return "", errors.Wrap(err, "get namespace")
+	}
+
+	return fmt.Sprintf("%s.%s.%s", hostname, svcName, namespace), nil
+}
+
+func getSecret(username apiv1alpha1.SystemUser) (string, error) {
+	path := filepath.Join(mysql.CredsMountPath, string(username))
+	sBytes, err := ioutil.ReadFile(path)
+	if err != nil {
+		return "", errors.Wrapf(err, "read %s", path)
+	}
+
+	return strings.TrimSpace(string(sBytes)), nil
+}
+
+func getPodIP(hostname string) (string, error) {
+	addrs, err := net.LookupHost(hostname)
+	if err != nil {
+		return "", errors.Wrapf(err, "lookup %s", hostname)
+	}
+	log.Println("lookup", hostname, addrs)
+
+	return addrs[0], nil
+}
+
+func lookup(svcName string) (sets.String, error) {
+	endpoints := sets.NewString()
+	_, srvRecords, err := net.LookupSRV("", "", svcName)
+	if err != nil {
+		return endpoints, err
+	}
+	for _, srvRecord := range srvRecords {
+		// The SRV records have the pattern $HOSTNAME.$SERVICE.$.NAMESPACE.svc.$CLUSTER_DNS_SUFFIX
+		// We only want $HOSTNAME.$SERVICE.$NAMESPACE because in the `selectDonor` function we
+		// compare the list generated here with the output of the `getFQDN` function
+		srv := strings.Split(srvRecord.Target, ".")
+		ep := strings.Join(srv[:3], ".")
+		endpoints.Insert(ep)
+	}
+	return endpoints, nil
+}
+
+func lockExists() (bool, error) {
+	_, err := os.Stat("/var/lib/mysql/bootstrap.lock")
+	if err != nil {
+		if os.IsNotExist(err) {
+			return false, nil
+		}
+		return false, errors.Wrap(err, "os stat")
+	}
+	return true, nil
+}
+
+func waitLockRemoval() error {
+	for {
+		exists, err := lockExists()
+		if err != nil {
+			return err
+		}
+		time.Sleep(time.Second)
+		if !exists {
+			return nil
+		}
+	}
+}
